Restructure service backends to own component logic
