+++ docs/implementations/contract-drafter/observations.md
@@ -2,10 +2,10 @@
 
 The dc43 contract drafter operates on runtime-agnostic observations. The
 core helper lives in
-`dc43.lib.components.contract_drafter.observations.draft_from_observations`
+`dc43.services.contracts.backend.drafting.draft_from_observations`
 and expects a schema snapshot plus optional metric payload produced by
 the data-quality engine. Use
-`dc43.lib.components.data_quality.integration.schema_snapshot` (or an
+`dc43.services.data_quality.backend.integration.schema_snapshot` (or an
 equivalent helper in your runtime) to gather the fields before delegating
 to the drafter.
 
@@ -27,8 +27,8 @@ The helper consumes three categories of inputs:
 ## Usage
 
 ```python
-from dc43.lib.components.contract_drafter.observations import draft_from_observations
-from dc43.lib.components.data_quality.integration import schema_snapshot
+from dc43.services.contracts.backend.drafting import draft_from_observations
+from dc43.services.data_quality.backend.integration import schema_snapshot
 
 schema = schema_snapshot(dataframe)
 metrics = validation_result.metrics
