+++ docs/implementations/integration/spark-dlt.md
@@ -5,7 +5,7 @@ dc43 keeps governance logic decoupled from runtime execution. The integration la
 ## Responsibilities
 
 1. **Resolve runtime identifiers** (paths, tables, dataset versions) and map them to contract ids.
-2. **Validate and coerce data** using helpers from `dc43.lib.components.data_quality.integration` while respecting enforcement flags.
+2. **Validate and coerce data** using helpers from `dc43.services.data_quality.backend.integration` while respecting enforcement flags.
 3. **Bridge runtime metrics** to the governance service so it can evaluate observations, record activity, and propose drafts when mismatches occur.
 4. **Expose ergonomic APIs** for pipelines (`read_with_contract`, `write_with_contract`, `expectations_from_contract`).
 
@@ -27,7 +27,7 @@ The canonical implementation lives in [`src/dc43/integration`](../../src/dc43/in
 
 * `spark_io.py` — High-level `read_with_contract` and `write_with_contract` wrappers for Spark DataFrames along with dataset resolution helpers.
 * `dlt_helpers.py` — Functions to translate ODCS expectations into Delta Live Tables expectations.
-* [`dc43.lib.components.data_quality.integration`](../../src/dc43/components/data_quality/integration/__init__.py) — Schema snapshots, expectation predicates, and metric builders used by adapters.
+* [`dc43.services.data_quality.backend.integration`](../../src/dc43/services/data_quality/backend/integration/__init__.py) — Schema snapshots, expectation predicates, and metric builders used by adapters.
 * [`dc43.services.governance`](../../src/dc43/services/governance) — Coordination service that links contracts, evaluates observations, and persists drafts.
 
 Pipelines typically import these helpers directly:
