+++ src/dc43/lib/components/contract_drafter/__init__.py
@@ -1,5 +1,13 @@
-"""Helpers that turn runtime feedback into ODCS drafts."""
+"""Compatibility exports for contract drafting helpers.
 
-from .observations import draft_from_observations, draft_from_validation_result
+The drafting logic now lives under :mod:`dc43.services.contracts.backend`.
+This module re-exports those helpers so library callers keep functioning while
+migrations take place.
+"""
+
+from dc43.services.contracts.backend.drafting import (
+    draft_from_observations,
+    draft_from_validation_result,
+)
 
 __all__ = ["draft_from_observations", "draft_from_validation_result"]
