+++ src/dc43/lib/components/contract_store/__init__.py
@@ -1,6 +1,6 @@
 """Contract store interfaces, gateways, and reference implementations."""
 
-from .interface import ContractStore
+from dc43.services.contracts.backend.stores.interface import ContractStore
 from .impl.filesystem import FSContractStore
 from .impl.delta import DeltaContractStore
 from .impl.collibra import (
