+++ src/dc43/services/contracts/backend/stores/delta.py
@@ -9,7 +9,7 @@ try:  # pragma: no cover - optional dependency
 except Exception:  # pragma: no cover
     SparkSession = object  # type: ignore
 
-from dc43.lib.components.contract_store.interface import ContractStore
+from .interface import ContractStore
 from dc43.odcs import as_odcs_dict, ensure_version, contract_identity, fingerprint, to_model
 from open_data_contract_standard.model import OpenDataContractStandard  # type: ignore
