+++ src/dc43/services/data_quality/__init__.py
@@ -1,8 +1,16 @@
 """Backend and client helpers for running data-quality services."""
 
-from dc43.lib.data_quality import ObservationPayload
-
-from .backend import DataQualityServiceBackend, LocalDataQualityServiceBackend
+from .backend import (
+    DataQualityManager,
+    DataQualityServiceBackend,
+    ExpectationSpec,
+    LocalDataQualityServiceBackend,
+    ObservationPayload,
+    ValidationResult,
+    apply_contract,
+    evaluate_contract,
+    expectation_specs,
+)
 from .client import DataQualityServiceClient, LocalDataQualityServiceClient
 
 __all__ = [
@@ -11,4 +19,10 @@ __all__ = [
     "DataQualityServiceClient",
     "LocalDataQualityServiceClient",
     "ObservationPayload",
+    "DataQualityManager",
+    "ValidationResult",
+    "ExpectationSpec",
+    "evaluate_contract",
+    "expectation_specs",
+    "apply_contract",
 ]
