+++ src/dc43/services/data_quality/models.py
@@ -1,5 +1,18 @@
-"""Backwards compatibility layer for :mod:`dc43.lib.data_quality`."""
+"""Shared data structures used across data-quality services."""
 
-from dc43.lib.data_quality import ObservationPayload
+from __future__ import annotations
+
+from dataclasses import dataclass
+from typing import Mapping, Optional
+
+
+@dataclass(slots=True)
+class ObservationPayload:
+    """Container describing cached observations for a dataset evaluation."""
+
+    metrics: Mapping[str, object]
+    schema: Optional[Mapping[str, Mapping[str, object]]] = None
+    reused: bool = False
+
 
 __all__ = ["ObservationPayload"]
