+++ src/dc43/services/governance/__init__.py
@@ -1,6 +1,11 @@
 """Governance orchestration service abstractions and local stubs."""
 
-from .backend import GovernanceServiceBackend, LocalGovernanceServiceBackend
+from .backend import (
+    DQClient,
+    DQStatus,
+    GovernanceServiceBackend,
+    LocalGovernanceServiceBackend,
+)
 from .client import (
     GovernanceServiceClient,
     LocalGovernanceServiceClient,
@@ -30,4 +35,6 @@ __all__ = [
     "LocalGovernanceServiceClient",
     "LocalGovernanceService",
     "build_local_governance_service",
+    "DQClient",
+    "DQStatus",
 ]
