+++ src/dc43/services/governance/backend/__init__.py
@@ -1,6 +1,7 @@
 """Backend contracts and stubs for governance services."""
 
+from .dq import DQClient, DQStatus
 from .interface import GovernanceServiceBackend
 from .local import LocalGovernanceServiceBackend
 
-__all__ = ["GovernanceServiceBackend", "LocalGovernanceServiceBackend"]
+__all__ = ["GovernanceServiceBackend", "LocalGovernanceServiceBackend", "DQClient", "DQStatus"]
