+++ src/dc43/services/governance/backend/local.py
@@ -6,16 +6,16 @@ from typing import Any, Callable, Dict, Mapping, Optional, Sequence
 
 from open_data_contract_standard.model import OpenDataContractStandard  # type: ignore
 
-from dc43.lib.components.contract_drafter import draft_from_validation_result
-from dc43.lib.components.contract_store.interface import ContractStore
-from dc43.lib.components.contract_validation import ValidationResult
-from dc43.lib.components.data_quality.governance import DQStatus
+from dc43.services.contracts.backend.drafting import draft_from_validation_result
+from dc43.services.contracts.backend.stores.interface import ContractStore
+from dc43.services.data_quality.backend.engine import ValidationResult
+from dc43.services.governance.backend.dq import DQStatus
 from dc43.odcs import contract_identity
 from dc43.services.contracts.backend import ContractServiceBackend
 from dc43.services.contracts.client import ContractServiceClient
 from dc43.services.data_quality.backend import DataQualityServiceBackend
 from dc43.services.data_quality.client import DataQualityServiceClient
-from dc43.lib.data_quality import ObservationPayload
+from dc43.services.data_quality.models import ObservationPayload
 
 from .interface import GovernanceServiceBackend
 from ..models import (
